SUPERNOVA: Divide rooms into 2 files
