+++ engines/supernova/state.h
@@ -28,6 +28,7 @@
 #include "common/rect.h"
 #include "common/keyboard.h"
 #include "supernova/rooms.h"
+#include "supernova/room.h"
 #include "supernova/sound.h"
 #include "supernova/game-manager.h"
